% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.sivs.R
\name{is.sivs}
\alias{is.sivs}
\title{Validate sivs Object}
\usage{
is.sivs(object)
}
\arguments{
\item{object}{Idealy the object that is produced by the sivs function.}
}
\value{
This function will return TRUE if it detects the function is truely a
sivs object, otherwise it will return FALSE.
}
\description{
A function to validate if the given object is truely from sivs function
}
