% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.sivs.R
\name{coef.sivs}
\alias{coef.sivs}
\title{Extract Coefficients from sivs object}
\usage{
\method{coef}{sivs}(object, step = "iterative.res")
}
\arguments{
\item{object}{An object of class "sivs"}

\item{step}{A a character string of length 1. It should either specify the
step ("iterative.res" or "rfe"), or step$subsetp (e.g "rfe$baseline").}
}
\description{
A function to extract the coefficients of "iterative.res" step
or any part of "rfe" such as "sivs_object$rfe$baseline" from a sivs object.
}
\examples{
\dontrun{
# getting the coefficients of features for the baseline runs in rfe
coef(object = sivs_object, step = "rfe$baseline")
}

}
