% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sivs.R
\name{plot.sivs}
\alias{plot.sivs}
\title{A plotting funcion for sivs object}
\usage{
\method{plot}{sivs}(
  object,
  type = c("frequency", "coef", "rfe"),
  suggestion_strictness = c(0.01, 0.05),
  ...
)
}
\arguments{
\item{object}{The object that is produced by the sivs function.}

\item{type}{Which plot do you want to have. Acceptable values are "frequency" and "coef" at the moment.}

\item{suggestion_strictness}{The strictness value that indicates how much the thresholds should be strict. For more details visit help page of `suggest()` function. If you want ru suppress the suggestion on the plot, set this to NULL.}

\item{...}{The other argument you might want to pass to each plot.}
}
\description{
A function to plot the object of the sivs function.
}
\details{
The rfe plot ignores the existance of intercept since intercept
cannot be removed in the recursive feature elimination step. This
is the reason that the number of features with vimp greater than 0
are differnt from the presented "Number of Features Left" in the
rfe plot by 1.
}
\examples{
\dontrun{
# to see all plots
layout(mat = matrix(c(1,2,3,3), nrow = 2, byrow = T))
plot(object = sivs_object)
layout(1)

# to plot only the Recursive Feature Elimination (rfe) results
plot(object = sivs_object, type = "rfe")

# suppress suggestion on rfe plot
plot(object = sivs_object, type = "rfe", suggestion_strictness = NULL)
}

}
